Add core Upgrade() function for upgrading templates (#468)
This is the heart of the upgrade logic (finally!) in
`common/upgrade/upgrade.go`. It's still missing many important parts,
but the basic skeleton is now here.

Also factor out the "prompt" flag into the common package; it had been
copy-pasted in two places and we're adding a third.

I'm using the convention `TODO(upgrade)` to note every place where more
work is needed for the template upgrade feature.