Add individual file hashing function (#485)
This is part of the template upgrades feature. I'm committing in pieces
because the entire feature is too big for one PR.

Why are there no tests? I've chosen to test this at the level of the
exported API. There are tests already written that will be sent out soon
that cover all of the code paths in this PR.