feat: detect previous unmerged files and abort upgrade (#499)
This is similar to how git won't let you proceed until you resolve merge
conflicts. When an upgrade operation begins, we look for any
`*.abcmerge_*` files that would indicate that there was a previous
upgrade operation that still needs to be manually resolved. If we were
to ignore this and upgrade anyway, the user would likely be incredibly
confused, and that would be bad.