Return merge conflicts and explanations from Upgrade() (#501)
The Upgrade() function now returns a `Result` struct, which explains
what happened.

If the Upgrade was a noop because there was no new template version,
then `Result.AlreadyUpToDate` is set.

If there were any merge conflicts, then the files that need to be
resolved are in `Result.Conflicts`, along with explanations. All other
non-conflicting files are in `Result.NonConflicts`.