Always allow un-clean git workspaces (#505)
Originally, I wanted to avoid saying "this template was installed from
SHA x" in the case where the git workspace had pending changes. The
pending changes might have been significant to the template output,
meaning that we weren't really installing that version of the template.

Now, I've given up on that. We will very often be installing from a
dirty workspace, such as when multiple templates are upgraded at once,
and the template directory is in the same repo as the output dir (like
in terraform GCP org repos). So we'll just keep it simple and accept a
little risk of confusion.