fix: template_location was mistakenly made into a required field
template_location is required for upgrading (so we add a new error case
to verify that), but it should not be required when unmarshaling a
manifest model. That unmarshaling validation has now been removed. It's
valid for a template to be rendered from a non-canonical template
location, in which case template_location will be empty string.