fix: include-from-destination bit should be reset on re-include
There was a bug here that we haven't actually hit, but could be
important for the template upgrade logic that will build on top of this.
The bug could cause files to be considered "included from destination"
even when they're not. The bug arises when the same path is included
multiple times with a mix of regular include and "include from
destination."

The fix is just to reset the state for "included from destination" when
a path is re-included.