Fix "include from destination" behavior in the upgrade case
This is kind of a bug, but the upgrade code isn't exposed yet, so not
really. In the case where an upgrade operation executes an `include`
action with `from: destination`, the file should be copied from the
ultimate user-visible destination folder, not the temp dir that we're
rendering into (which will just be empty).