Redesign the "run" package interface
Since we're accumulating more use cases for running external commands,
it's time to clean up this interface. I'm about to add some new code
that runs the "patch" command, which requires passing standard input,
which we haven't done before.

So it's time to add a nice "functional options" interface for the Run()
function that allows options to be mixed-and-matched.