Include patches in manifest file when rendering (#530)
In the case where a file is included with `from: destination`, which is
used to modify existing files in place, we'll save a diff when
overwriting the destination file when commiting the rendered template
output. This allows us to reverse the change later, which is important
for the template upgrade process.