Reorganize processGlobs and behavior on include path not existing
This is functionally the same as before, but we move some edge case handling. Before, if processGlobs didnt find the file it was looking for, it would return an error. Now, that "no such file" logic has been moved into each of the two *callers* of processGlobs. This sets up up or a future PR where we can attempt processGlobs() on multiple directories and fail only if none of them are successful.