Remember source directory for "include from destination" (#539)
In this edge case:

 - We're doing an upgrade
 - The template has an include action with `from: destination`
- Some upgrade-reversal patches were successfully applied automatically
by abc
- Other upgrade-reversal patches had merge conflicts and were resolved
manually by the user
 
... then there's some ambiguity about which directory is the
"destination" directory (in the sense of "include from destination").
There are two directories that could be considered the be the source
directory of the "include from destination" action. The first one is the
real user-visible directory where the template is installed, and the
other is the temp directory where the upgrade command saves its
successful upgrade-reversal-patched files.

Therefore, whenever we do an "include from destination" action, we save
the directory that the file came from. This will be needed later in
upgrade.go when we merge the new template output with the installed
template directory.