Return a result struct from Render() (#540)
Before now, we were hackily relying on the caller of Render() to go and
find the manifest file themself. Now we return the manifest path.

We also return the set of files that were included from destination, and
the directory from which each file was copied. This will be needed for
the upgrade operation soon.