Fix a bug in golden-test verify for templates without tests
Recently we added the ability for `golden-test verify` and `record` to
crawl the directory tree looking for tests to process. This failed in
the case where it found a template with no tests (it's valid for a
template to not have tests).

The crawling logic was untested, so we add tests now.