Implement include-from-destination patch reversal (#547)
This is part of supporting "included-from-destination" files in the
upgrade algorithm. The basic algorithm in this:
- When we modify a preexisting file in-place using the
include-from-destination mechanism, we save a patch in the manifest file
that will undo our modification. This happens before the upgrade.
- When it's time to upgrade, we apply the patch that was saved in the
manifest to undo the previous change. Then, when we render the upgraded
template, it sees the old version of the file before it was modified by
the old template version.

For example, suppose we have an outputs.tf file. When we initially
render the template, we add an `output {...}` to outputs.tf, and we save
a patch in the manifest YAML file that will remove this output in the
future. When it's time to upgrade, we apply the patch to remove the
`output {...}` from outputs.tf that was added in the past. The new
template version is then rendered, and it can either re-add the same
`output {...}` block, perhaps with changes (like a new name), or perhaps
identically to what was there before, depending on what the new template
specifies.

This isn't fully complete yet. The following PR will include adjustments
to the merge algorithm to account for the special case of patch-reversed
files, and include a bunch of tests.