Merge algorithm adjustments for patched include-from-destination files (#548)
This builds on #547. 

This PR does a few things:
- Handles patch-reversed-included-from-destination files specially in
the merge algorithm. Since these files have already been patched
successfully or resolved by the user, they never constitute a conflict
in the merge phase. Essentially the conflict has already been resolved.
- Adds the temp dir containing patch-reversed files as a source
directory for the include action. This is necessary for the "include"
action to see the patch-reversed file instead of the unpatched file.
 - Add a bunch of tests for upgrading with patch reversal.