Handle patch-reversal conflict in the upgrade command (#571)
When the upgrade operation fails because one of the patches in the
manifest couldn't be applied, then we'll ask the user to manually apply
the failed hunks. This PR adds the instructions to the user from the
upgrade command. This includes giving the user the exact command to run
after they resolve the conflict.