Return DownloadMetadata from Upgrade() (#572)
In an upcoming PR, we need to distinguish between remote templates (e.g.
from GitHub) and local templates (e.g. in some local directory). This
distinction matters for the "upgrade all templates under this directory"
feature, for complex reasons that probably aren't worth going into here
(feel free to ask if you're interested).

As part of this, we change the DownloadMetadata.LocationType field from
a string to an enum, because we want to add an IsRemote() method to
LocationType. And also because this is an exported API now, and that
suggests being cleaner with types.