Add "upgrade all" feature (#580)
With this PR, users can just run "abc upgrade" and we'll find all the
template installations under the current directory and upgrade all of
them. The user no longer has to point to a manifest file, or even know
what a manifest file is.

It's still missing some tests, which I'll add in a future PR. This one
is already huge.

This is alpha quality at best. It probably has many critical bugs. But
I've been working on this for months and it's time to get something
committed.