fix: --manifest-only was failing if unrelated files exist in the dest
In the just submitted PR #582, there was a bug. The bug was that if the
template outputted any files that already existed in the destination,
the render operation would fail. However, in --manifest-only mode, this
is fine and expected. These files *should* already exist.

Renamed CopyParams.Overwrite to CopyParams.AllowExisting to reflect that
it's now used in situations where an overwrite is not actually
occurring; now it's just a flag to "not fail if the file exists in the
destination, not necessarily implying it will be overwritten."