fix: "upgrade all" should halt on merge conflict
When a template fails to upgrade due to conflicts between the template
and local edits, we should stop upgrading any further templates until the merge conflicts are resolved.

It used to work this way, but broke in a refactor by accident.

Added a test to prevent regressing on this behavior again.