fix: --manifest-only was failing if unrelated files exist in the dest (#583)
In the just submitted PR #582, there was a bug: if the template
outputted any files that already existed in the destination, the render
operation would fail. However, in --manifest-only mode, this is fine and
expected. These files *should* already exist.

Renamed CopyParams.Overwrite to CopyParams.AllowExisting to reflect that
it's now used in situations where an overwrite is not actually
occurring; now it's just a flag to indicate "don't fail if the file
exists in the destination, which doesn't necessarily imply that it will
be overwritten."