Adjust the suggested commands to continue an upgrade (#585)
When the upgrade operation is interrupted due to a merge conflict that
needs to be manually resolved, we instruct the user how to continue the
upgrade by running a command. Before this PR, we attempted to print the
exact command to run to continue the upgrade. But I realized that this
approach was flawed because we need to print all the flag values that
might be needed (like --version). These flags may or may not have been
provided by the user in their upgrade command.

That would have been too complex. So the new solution in this PR is to
tell the user to re-run their previous upgrade command with a certain
flag added, without attempting to provide the full set of flags.