Implement the upgrade --resume-from feature
When there is a "patch reversal conflict" (which happens when there's a
merge conflict during upgrade on a file that was modified in place),
then the user must run `abc upgrade --resume-from=some/manifest.yaml` to
resume the upgrade. This PR implements the feature; the flag existed
before but was unused.