Disallow `..` paths in manifest files (#593)
In a manifest file, we store a list of output files from a template,
along with their relative path from the template output directory. No
`..` directory traversals should appear in this file. If they do appear,
it could cause abc to touch files that it shouldn't be touching. So
detect and disallow this.