Finish implementing --resume-from/--already-resolved
This code runs when recovering from merge conflicts. After a merge
conflict is resolved, the user runs `abc upgrade <directory>
--resume-from=some-manifest.yaml --already-resolved=<list-of-files>`.
In the case where there are multiple template installations to be
upgraded, the `--already-resolved` flag should only apply to the first
one, and later upgrades should be attempted from scratch.

Most of this is just test updates.