Allow templates to be upgraded even when installed from a random local dir (#597)
If a template is rendered from a "canonical" location, such as a github
repo, then we store the source of template in the manifest. Future
template upgrades will check that location for new versions of the
template. But when the template is installed from a non-canonical
location, like `abc render ~/alice-silly-temp-dir`, then there's no
location stored in the manifest to be checked for updates.

This PR adds the "--template-location" flag to the abc upgrade command
to provide a template location to upgrade to, overriding the manifest.
So somebody else can come along later and do something like `abc upgrade
--template-location github.com/me/myrepo/mytemplate` or `abc upgrade
--template-location ~/bob-silly-temp-dir`, even if the manifest doesn't
have a stored location.