For certain conflict types, leave local filenames unchanged
Before this PR, when there was a conflict between a local edit and an
edited file in the upstream template, we'd rename the user's local file
to something like `foo.abcmerge_locally_edited` and drop the template's
edited version of the file as `foo.abcmerge_new_from_template`.

However, after using `abc` as a user, I think a better user experience
is to leave the user's filename unchanged. The sensible workflow for a
user who's facing a merge conflict is to diff their local file with the
incoming file and selectively incorporate changes from the upstream
template file. These files don't have equal standing from the user's
point of view: the user wants to keep what they have and carefully vet
and integrate the upstream changes. This PR makes that change.