fix: ignored inputs should not appear in manifest (#599)
There was a bug in the way unknown inputs were handled. They would be
correctly ignored for the purposes of rendering the template, but would
still appear in the manifest. This didn't cause serious problems, but
was confusing and gross; a manifest should not save any inputs that
weren't actually passed to the template during rendering.