Replace forbidden characters in manifest filenames with underscore (#600)
Before this PR, we URL-escaped the template location and used that in
the manifest filename. This was ugly and unpleasant to deal with. The
`/` character often appears in these names, which means our manifest
filenames were riddled with `%2F` escape sequences which are unpleasant
to type, tab-complete, and look at.

The new scheme is to have an allowlist of characters that can appear in
manifest filenames, and replace any other character with an underscore.
This might seem less safe, since there could be collisions, but it's not
a problem because the manifest filename also incorporates a nanosecond
timestamp which will prevent collisions.