Add --accept-defaults flag, and error when user intent is unclear (#604)
This avoids a specific poor user experience. Suppose the user runs `abc
upgrade` (without --prompt), which is a very reasonable thing to do.
Suppose the upgraded version of the template has a new input with a
default value for that input. It would be bad for abc to just silently
use the default value for that new input; technically you could argue
that the user should have used the --prompt flag if they wanted a chance
to override the default, but that would be a footgun and we can't expect
users to be that diligent. So we'll reject the current operation and ask
the user to clarify their intent with either --prompt or
--accept-defaults.