Handle edited-but-equal merge conflict more intelligently
In the case where:

 - A template outputs a file
 - The user edits that file
 - We upgrade to a template that also changes that same file to
   the exact same contents as the user's local modified copy

There should be no conflict, because the locally edited copy and the new
template "agree" on the desired contents of the file.

This was observed when testing upgrades in infra-gcp and finding
spurious merge conflicts asking me to resolve files that were identical.

Also, add a file Close() that was missing 🙃