Rename --manifest-only to --backfill-manifest-only (#617)
The new naming is slightly more clear. I recently found an edge case in
template upgrades that modify files in place, which will be addressed in
an upcoming PR. Before that though, I wanted to make it clear this
operation is slightly different than a normal render, it's a different
verb, "backfill".

Alternative considered: make `abc backfill` its own command. This would
be too heavyweight of a solution for a small problem.

This PR is just a trivial rename with no logic changes.