Add the --continue-without-patches flag for `render` (#619)
There's a funky edge case in the render command. In the case where:
1. the user is backfilling a manifest for a previously rendered template
and
2. the template being rendered uses the "include from destination"
feature

In this case we can't create a fully accurate manifest. For files that
are "included from destination" (aka modified in place), the manifest is
supposed to store a patch that will undo the changes that were made
while rendering the template. But this isn't possible when backfilling a
manifest for an already-rendered template, because the previous contents
of the file (before it was modified in place) are no longer available.

In this PR, we detect this case and explain the problem to the user. If
the user is OK with creating a manifest that missing these patches, they
can re-run the upgrade command with the flag
`--continue-without-patches`.

I don't see any cleaner solution; fundamentally the
include-from-destination feature is kinda incompatible with backfilling
manifests for old templates.