Remove Manifest option in favor of SkipManifest
Going forward, we want render operations to default to producing a
manifest. Therefore, in the Render() library API, we want to change to
sense of the Manifest bool to SkipManifest so that the default value of
false does the desirable default behavior.

This is technically part of an exported API that shouldn't be changed,
but since there's only one non-CLI user of this, and he's on our team,
we'll allow it to change.

The command line flag also changes from --manifest to --skip-manifest.
This is also a backwards-incompatible change, but since the flag is
documented as experimental, we'll allow it.

We also update a bunch of tests to assert on the produced manifests.