fix: go.mod file should be included in template output (#628)
The hello_jupiter file contained a go.mod file that was supposed to part
of the template output, but it was never mentioned in the spec file and
therefore was not actually used.

The fix is to include all files by wildcard. This requires updating the
api_version to a version where wildcards are supported (v1beta6 is the
latest).

We also ran `abc golden-test record` to update the recorded golden-test
output with the new file.