Create target directory in CopyFile (#627)
There have been two bugs now where we tried to CopyFile to a directory
that doesn't exist. For all the use cases in abc, we always want to
create the directory if it doesn't exist, so we just add that to
CopyFile.

Also, CopyFile was missing unit tests so this PR adds them.