fix: the manifest upgrade_channel field was sometimes omitted (#629)
There was a bug in the case where we do `abc upgrade --version foo` to
upgrade (or downgrade) to a specific version. In that case, the
`upgrade_channel` field in the new manifest would be empty. The desired
behavior is for the new manifest to have an upgrade_channel field equal
to --upgrade-channel (if provided) or just inherit the upgrade_channel
field from the old manifest.

The `FlagUpgradeChannel` struct field is renamed to `UpgradeChannel`
because now it doesn't always come from a flag.