Refactor manifest loading in UpgradeAll() (#632)
This is an uninteresting refactor of the code that loads the manifests
to be upgraded and sorts them. The main idea is that we're taking code
out of UpgradeAll() and moving it into a new subroutine,
`manifestsToUpgrade()`.

This is to prepare for an upcoming feature that adds filtering of
manifests to be upgraded.