Return AlreadyUpToDate if the user inputs haven't changed
The old behavior was to noop in the case where the upstream template
hasn't changed.

The new behavior is to noop in the case where the upstream template
hasn't changed, AND the inputs to the template haven't changed. This
includes all possible sources of inputs.

Of course this noop'ing can be overwritten with --continue-if-current to
still force the upgrade operation to run.